[TASK] added the basic frontend and backend
Added a frontend, that can display the users cam and can take still images.

Added a basic infrastruktor for different output and input devices in the backend.

Added  functionality to the frontent that lets the user add a LED to the backend and change if its turned on or off.